Add contentHash to bundle names, enable hashed module ids (HashedModuleIdsPlugin)
Configuration options to include a bundle's hash in its name, remove runtime from the bundle's hash and save it as a single chunk (runtimeChunk: 'single'), and use HashedModuleIdsPlugin so that vendor bundles do not change their hash when unrelated files are updated (as seen here: https://webpack.js.org/guides/caching/). This is necessary because module ids are changed if dependencies are loaded in a different order, and because bundle hashes use module ids.