Add async for getting Leetcode data with content script Popup.js will now ask the content script to get Leetcode page data when on the leetcode page if the user has requested to pair on their problem
Content.js will now start looking for data on the page when prompted by popup.js. Content.js will send a message to popup.js notifying whether page data was found or not.

Popup.js will start the app when content script confirms it has page data or notify the user to refresh the page if page data was not found.